WIP: Worked on matherep-reveal
reveal.js can do things a bit easier and has many ready to use plugins and
elements. It is a bit more 'normal' (like a classic presentation) but
not less fancy. I decided to use reveal.js instead of impress.js,
for several reasons: among other things I like the presenters plugin and
the ready to use themes and effects.
I copied the content of matherep-impress to matherep-reveal